+++ packages/dialtone-vue2/components/popover/popover.vue
@@ -68,6 +68,7 @@
           :content-class="headerClass"
           type="header"
           :show-close-button="showCloseButton"
+          :close-button-props="closeButtonProps"
           @close="closePopover"
         >
           <template #content>
@@ -227,6 +228,15 @@ export default {
     ariaLabel: {
       type: String,
       default: null,
+    },
+
+    /**
+     * A set of props to be passed into the popover's header close button.
+     * Requires an 'ariaLabel' property, when the header popover is visible
+     */
+    closeButtonProps: {
+      type: Object,
+      default: () => ({}),
     },
 
     /**