+++ packages/dialtone-vue2/recipes/buttons/callbar_button_with_dropdown/callbar_button_with_dropdown.vue
@@ -28,6 +28,7 @@
     <dt-dropdown
       v-if="showArrowButton"
       :id="id"
+      :modal="false"
       :fallback-placements="fallbackPlacements"
       :open="open"
       :placement="placement"